requirements.txt
streamlit
python-docx
cairosvg
PyPDF2
